Fixes for second doc
Note: this splits out results between q1-yes and q1-no-in-nj by duplicating the large result set.  i think doing this in a declarative way as i did here makes the most sense because it means we don't have to make any major changes to the schema or code to get the conditionality that we want here. however, this does come at the cost of a lot of duplication