Add NJ careers for Q1=laid-off-in-nj
note: right now it was simpler to add nj-careers at the bottom immediately upon selecting one of the Q2 options that results in the A-M list for Q3 because otherwise we would need to split the A-M definitions. a nice way to fix this might be to build a way to handle conditionality within definitions but that seems tough. a better way might be to abstract the definitions of the A-M questions from the rendering of the final benefits/protections/etc results